great

+++ Cluster/Cluster_experiments.R
@@ -1,14 +1,25 @@
 
 library(batchtools)
-
+setwd("H:/Documents/MdS/Cluster")
 
 set.seed <- 42
 
 # Registry ----------------------------------------------------------------
-reg_name <- "example_experiments"
-reg_dir <- file.path("registries", reg_name)
-unlink(reg_dir, recursive = TRUE)
-makeExperimentRegistry(file.dir = reg_dir)
+
+### Setting up the repository
+start_from_scratch <- T # if true, removes all repository and creates a new one
+
+reg_name <- "MdS_registry"
+reg_dir <- sprintf("%s/registries/%s", getwd(), reg_name)
+
+if (start_from_scratch) {
+  dir.create("registries", showWarnings = FALSE)
+  unlink(reg_dir, recursive = TRUE)
+  reg <- makeExperimentRegistry(file.dir = reg_dir)
+} else {
+  reg <- loadRegistry(file.dir = reg_dir, writeable = TRUE)
+}
+
 
 # Problems -----------------------------------------------------------
 source("problems.R")
@@ -18,46 +29,36 @@ addProblem(name = "sim_graphs", fun = sim_graphs, seed = 43)
 # Algorithms -----------------------------------------------------------
 source("Algorithms.R")
 addAlgorithm(name = "MdS", fun = MdS_wrapper)
-addAlgorithm(name = "JGL", fun = jgl_wrapper)
+addAlgorithm(name = "JGL", fun = JGL_wrapper)
 addAlgorithm(name = "glasso", fun = glasso_wrapper)
 
 # Experiments -----------------------------------------------------------
-prob_design <- list(sim_graphs = expand.grid(p = 150,
-                                           perc1 = 0.05,
-                                           perc2 = 0.05,
-                                           length_var1 = 3,
-                                           length_var2 = 3,
-                                           observes = 52))
+prob_design <- list(sim_graphs = expand.grid(p = 12,
+                                           perc1 = 0.,
+                                           perc2 = 0.2,
+                                           length_var1 = 2,
+                                           length_var2 = 2,
+                                           observes = 2000))
 
 
 #Setup the weight matrix
-th11 <- c(0,0,0,0,0,0,0,0,0) #1
-th12 <- c(1,0,0,0,0,0,0,0,0) #2
-th13 <- c(1,1,0,0,0,0,0,0,0) #3
-th21 <- c(1,0,0,0,0,0,0,0,0) #4
-th22 <- c(0,1,0,1,0,0,0,0,0) #5
-th23 <- c(0,0,1,1,1,0,0,0,0) #6
-th31 <- c(1,0,0,1,0,0,0,0,0) #7
-th32 <- c(0,1,0,0,1,0,1,0,0) #8
-th33 <- c(0,0,1,0,0,1,1,1,0) #9
 
-w <- rbind(th11,th12, th13, th21, th22, th23, th31, th32, th33)
-colnames(w) <- rownames(w)
 
 ###############################
 
-#Lambdas for MdS
-lam1_MdS <- seq(0.01,15, length.out = 20)
-lam2_MdS <- seq(0.1,2, length.out = 8)
 
-#Lambdas for JGL
-lam1_JGL <- seq(0.02,2, length.out = 20)
-lam2_JGL <- seq(0.1,2, length.out = 8)
+# a min lambda1
+# b max lambda1
+# c min lambda2
+# d max lambda2
 
+# L1 step count for lambda1
+# L2 step count for lambda2
 
-algo_design <- list(MdS = expand.grid(lam1 = lam1_MdS, lam2 = lam2_MdS, w = w),
-                    JGL = expand.grid(lam1 = lam1_JGL, lam2 = lam2_JGL),
-                    glasso = expand.grid(lam1 = 1, lam2 = 1))
+algo_design <- list(MdS = expand.grid(a = 0.1, b = 30, c = 0.1, d = 2, L1 = 20, L2 = 8 ,
+                                      weight_matrix = c("horizontal-vertikal",  "full")),
+                    JGL = expand.grid(a = 0.1, b = 2, c = 0.1, d = 2, L1 = 20, L2 = 8 ),
+                    glasso = expand.grid())
 
 addExperiments(prob_design, algo_design, repls = 5) #100
 summarizeExperiments()
